Create a helper for transforming the Embed slice
