+++ content/webapp/services/prismic/types/body.ts
@@ -82,7 +82,7 @@ export type Embed = Slice<
   'embed',
   {
     embed: EmbedField;
-    caption: KeyTextField;
+    caption: RichTextField;
   }
 >;
